db: move download code to a separate file
